make sure /api/github does nothing with public projects
If a project is public we should run tests on travis and use that status
to deploy. However, we want chewie to know about the commits. That way
we are able to show status on the chewie index when we get that up and
running. This commit also adds a test to make sure deploy is not called.